Rework InputOptionsSerializer (#77)
In #28, the feature to forward the command options was introduced. It had to later be followed up by #41, #42 in order to fix the issue where the option value may require quoting. This piece of code is heavily under-tested (and it was really hard to test it as is) and is prone to bugs (I found 2 adding tests).

In this PR, I'm extracting this element into `InputOptionsSerializer` as a closed piece (static & final) as IMO there is no need to make it extensible: if broken it needs to be fixed. This extraction allows much easier testing.

This PR also fixes:

- the case of negatable options
- detection quoting with may fail to detect some special characters