Introduce a custom logger (#80)
A significant part of the code is about logging, which in the current state is both hard to test (even manually) and very unstable (it is the constant source of test failure).

This PR introduces a dedicated logger which we can later manage to inject and will allow to have better control of the parallel processing output during the tests as well as testing the output itself.

I also think the logger itself will need rework, but this is for later. This PR focuses on giving the exact same output.