Add an error handler (#88)
Back when we were confronted with the issue of error handling we settled for a quick and easy solution. This current solution has several flaws:

- It has some Container (i.e. incompatible with non Symfony app) dependent code
- It offers no convenient extension point
- Offers toggled global based behaviour which in this peculiar instance is not inherently bad but is not ideal neither

For this reason I'm introducing an error handler interface. This offers a better extension point and clearer code as well. The current implementation aims to stay in line with the existing behaviour.

Note that a bug has been fixed which is that current the Container was never reset even if resettable (due to `class_exists()` returning `false` – should be `interface_exists()`).