Make the logger as a runtime parameter (#90)
When working on a web app, it is common to declare the logger as a service registered to the container and use it like another service. For console commands however, this is not working as well. Indeed your logger may depend of the output for example, in which case it cannot be passed to the command as a regular service.

For this reason I am now passing the logger as a "runtime parameter" alongside the IO objects.