Pass progress symbol as a parameter instead of part of the constructor (#97)
This offers more flexibility as it makes the instantiation of the standard logger no longer dependent of the progress symbol. This means if we remove the progress symbol from being passed to the parallel executor (for example to use a default value), this will not be a problem (currently that default value _needs_ to be in Parallelization).