Introduce the ParallelExecutorFactory (#96)
This PR introduces the biggest piece that will allow to make it a lot easier to port the core of the trait to a regular command.

The `ParallelExecutorFactory` allows to create a `ParallelExecutor` with minimal code (i.e. leveraging the default values as much as possible) whilst still allowing the user to easily override any parameter.