Split the PHPStan configuration (#135)
I would like to go to higher levels for PHPStan. However the cost of increasing the level for the tests is way too high for my liking. Hence splitting the configuration in two would allow to increase the level for the source code without paying too high of a price for the tests.