Always pass the full IO (#167)
This more of a bigger BC break with the last beta since it changes the signature of 2 new methods, and one annoying BC break for when coming to 1.x.

The reason why I'm willing to introduce this change nonetheless is because I believe it is much more future-proof. Indeed although passing both the input & output may look useless at first glance, as soon as you are using `SymfonyStyle` for example, you will need both.