fix: Remove hooks API deprecations (#275)
This part of the API should not be deprecated. The idea of
`ParallelCommand` is to provide a leaner and alternative API, not to
replace the `Parallelization` trait.

A number of configuration methods are and should remain deprecated as
they can now be easily configured via
`configureParallelExecutableFactory()` instead*.

*: the interest here is not so much to provide a difference from a user
perspective, but rather to have one central point of customization, the
`ParallelExecutorFactory`. Otherwise if we add/change a setting there we
need to change it both in the factory and the trait.